refactor(code): rename files and add lines on readme
