+++ README.md
@@ -1,5 +1,26 @@
 # EstructuraDatos01_GrupoD_ProyectoFinal
 
 # Implementacion de los TADs vistos en el avance numero 1 
-# Codigo tiene como objetivo ser funcional como un juego de peleas y combos
+# Codigo tiene como objetivo ser funcional como un juego de 
+# peleas y combos
 # donde se guarda la informacion de los jugadores
+
+# HEADER FILES (.h)
+# used for all declarations in this file
+
+
+# IMPLEMENTATION FILES (.cpp) = src
+# used for definitions/implementations
+
+# when we wanna use functions declared inside .h we just include .h in .cpp file
+
+# we include .h's on .cpp files
+
+# 
+#
+#
+#
+#
+#
+#
+#
