+++ src/score.cpp
